Staff m c finished. minor issues still remain

+++ App/Admin/Models/Photo.php
@@ -51,25 +51,6 @@ class Photo extends \Core\Model
     public $valid_types = ["jpg","png","jpeg"];
 
 
-    /**
-     * Error messages for photo upload process
-     *
-     * @var array
-     */
-
-
-    public $upload_errors_array = [
-
-            UPLOAD_ERR_OK         => "There is no error",
-            UPLOAD_ERR_INI_SIZE   => "The upload file exceeds the upload_max_size directive",
-            UPLOAD_ERR_FORM_SIZE  => "The uploaded file exceeds the max_file_size directive",
-            UPLOAD_ERR_PARTIAL    => "The uploaded file was only partially uploaded",
-            UPLOAD_ERR_NO_FILE    => "No file was uploaded",
-            UPLOAD_ERR_NO_TMP_DIR => "Missing a temporery folder",
-            UPLOAD_ERR_CANT_WRITE => "Failed to write file to disk",
-            UPLOAD_ERR_EXTENSION  => "A PHP extension stopped to file upload"
-    ];
-
 
     /**
      * Class constructor
@@ -281,22 +262,25 @@ class Photo extends \Core\Model
 
     public function deleteById($id){
 
-        $this->deletePhoto($id);
+        if($this->deletePhoto($id)){
 
-        $sql = "DELETE FROM media WHERE media_id = :photo_id ";
+            $sql = "DELETE FROM media WHERE media_id = :photo_id ";
 
-        $db = static::getDB();
-
-
-        $stmt = $db->prepare($sql);
-
-        $stmt->bindParam(":photo_id",$id,PDO::PARAM_INT);
-
-        $stmt->setFetchMode();
-
-        $stmt->execute();
+            $db = static::getDB();
 
 
+            $stmt = $db->prepare($sql);
+
+            $stmt->bindParam(":photo_id",$id,PDO::PARAM_INT);
+
+            $stmt->setFetchMode();
+
+            $stmt->execute();
+
+            return true;
+        }
+
+        return false;
 
     }
